droplet: fix bad usage of strncpy() in utils.c
Previously dpl_set_addr_family() would use strncpy() to move the data
into the destination buffer. This produces quite a lot of issues, so we
replace it with memmove() and NULL-terminate the strings ourselves.
As a sideeffect, calling dpl_set_addr_family() with host == new_host is
not UB anymore.